+++ composition/src/utils/constants.ts
@@ -1,10 +1,11 @@
-import { DirectiveDefinitionNode, Kind, ScalarTypeDefinitionNode } from 'graphql';
+import { DEFAULT_DEPRECATION_REASON, DirectiveDefinitionNode, Kind, ScalarTypeDefinitionNode } from 'graphql';
 import { stringArrayToNameNodeArray, stringToNamedTypeNode, stringToNameNode } from '../ast/utils';
 import {
   ARGUMENT_DEFINITION_UPPER,
   AUTHENTICATED,
   BOOLEAN_SCALAR,
   COMPOSE_DIRECTIVE,
+  DEFAULT,
   DEPRECATED,
   ENUM_UPPER,
   ENUM_VALUE_UPPER,
@@ -37,7 +38,7 @@ import {
   SCOPE_SCALAR,
   SCOPES,
   SHAREABLE,
-  SOURCE_ID,
+  SOURCE_NAME,
   SPECIFIED_BY,
   STRING_SCALAR,
   TAG,
@@ -71,7 +72,7 @@ export const DEPRECATED_DEFINITION: MutableDirectiveDefinitionNode = {
       type: stringToNamedTypeNode(STRING_SCALAR),
       defaultValue: {
         kind: Kind.STRING,
-        value: 'No longer supported',
+        value: DEFAULT_DEPRECATION_REASON,
       },
     },
   ],
@@ -102,7 +103,7 @@ const EXTERNAL_DEFINITION: DirectiveDefinitionNode = {
   repeatable: false,
 };
 
-// directive @eventsPublish(topic: String!, sourceID: String) on FIELD_DEFINITION
+// directive @eventsPublish(topic: String!, sourceName: String! = "default") on FIELD_DEFINITION
 const EVENTS_PUBLISH_DEFINITION: DirectiveDefinitionNode = {
   arguments: [
     {
@@ -115,8 +116,15 @@ const EVENTS_PUBLISH_DEFINITION: DirectiveDefinitionNode = {
     },
     {
       kind: Kind.INPUT_VALUE_DEFINITION,
-      name: stringToNameNode(SOURCE_ID),
-      type: stringToNamedTypeNode(STRING_SCALAR),
+      name: stringToNameNode(SOURCE_NAME),
+      type: {
+        kind: Kind.NON_NULL_TYPE,
+        type: stringToNamedTypeNode(STRING_SCALAR),
+      },
+      defaultValue: {
+        kind: Kind.STRING,
+        value: DEFAULT,
+      },
     },
   ],
   kind: Kind.DIRECTIVE_DEFINITION,
@@ -125,7 +133,7 @@ const EVENTS_PUBLISH_DEFINITION: DirectiveDefinitionNode = {
   repeatable: false,
 };
 
-// directive @eventsRequest(topic: String!, sourceID: String) on FIELD_DEFINITION
+// directive @eventsRequest(topic: String!, sourceName: String! = "default") on FIELD_DEFINITION
 const EVENTS_REQUEST_DEFINITION: DirectiveDefinitionNode = {
   arguments: [
     {
@@ -138,8 +146,15 @@ const EVENTS_REQUEST_DEFINITION: DirectiveDefinitionNode = {
     },
     {
       kind: Kind.INPUT_VALUE_DEFINITION,
-      name: stringToNameNode(SOURCE_ID),
-      type: stringToNamedTypeNode(STRING_SCALAR),
+      name: stringToNameNode(SOURCE_NAME),
+      type: {
+        kind: Kind.NON_NULL_TYPE,
+        type: stringToNamedTypeNode(STRING_SCALAR),
+      },
+      defaultValue: {
+        kind: Kind.STRING,
+        value: DEFAULT,
+      },
     },
   ],
   kind: Kind.DIRECTIVE_DEFINITION,
@@ -148,7 +163,7 @@ const EVENTS_REQUEST_DEFINITION: DirectiveDefinitionNode = {
   repeatable: false,
 };
 
-// directive @eventsSubscribe(topic: String!, sourceID: String) on FIELD_DEFINITION
+// directive @eventsSubscribe(topic: String!, sourceName: String! = "default") on FIELD_DEFINITION
 const EVENTS_SUBSCRIBE_DEFINITION: DirectiveDefinitionNode = {
   arguments: [
     {
@@ -161,8 +176,15 @@ const EVENTS_SUBSCRIBE_DEFINITION: DirectiveDefinitionNode = {
     },
     {
       kind: Kind.INPUT_VALUE_DEFINITION,
-      name: stringToNameNode(SOURCE_ID),
-      type: stringToNamedTypeNode(STRING_SCALAR),
+      name: stringToNameNode(SOURCE_NAME),
+      type: {
+        kind: Kind.NON_NULL_TYPE,
+        type: stringToNamedTypeNode(STRING_SCALAR),
+      },
+      defaultValue: {
+        kind: Kind.STRING,
+        value: DEFAULT,
+      },
     },
   ],
   kind: Kind.DIRECTIVE_DEFINITION,
@@ -510,3 +532,27 @@ export const SCOPE_SCALAR_DEFINITION: MutableScalarNode = {
 export const MAXIMUM_TYPE_NESTING = 30;
 
 export const INHERITABLE_DIRECTIVE_NAMES = [EXTERNAL, SHAREABLE];
+
+export const baseDirectives = `
+  directive @deprecated(reason: String = "No longer supported") on ARGUMENT_DEFINITION | ENUM_VALUE | FIELD_DEFINITION | INPUT_FIELD_DEFINITION
+  directive @extends on INTERFACE | OBJECT
+  directive @external on FIELD_DEFINITION | OBJECT
+  directive @eventsPublish(topic: String!, sourceID: String) on FIELD_DEFINITION
+  directive @eventsRequest(topic: String!, sourceID: String) on FIELD_DEFINITION
+  directive @eventsSubscribe(topic: String!, sourceID: String) on FIELD_DEFINITION
+  directive @key(fields: openfed__FieldSet!, resolvable: Boolean = true) repeatable on INTERFACE | OBJECT
+  directive @provides(fields: openfed__FieldSet!) on FIELD_DEFINITION
+  directive @requires(fields: openfed__FieldSet!) on FIELD_DEFINITION
+  directive @specifiedBy(url: String!) on SCALAR
+  directive @tag(name: String!) repeatable on ARGUMENT_DEFINITION | ENUM | ENUM_VALUE | FIELD_DEFINITION | INPUT_FIELD_DEFINITION | INPUT_OBJECT | INTERFACE | OBJECT | SCALAR | UNION
+  directive @authenticated on ENUM | FIELD_DEFINITION | INTERFACE | OBJECT | SCALAR
+  directive @composeDirective(name: String!) repeatable on SCHEMA
+  directive @inaccessible on ARGUMENT_DEFINITION | ENUM | ENUM_VALUE | FIELD_DEFINITION | INPUT_FIELD_DEFINITION | INPUT_OBJECT | INTERFACE | OBJECT | SCALAR | UNION
+  directive @interfaceObject on OBJECT
+  directive @link(url: String!, as: String, for: String, import: [String]) repeatable on SCHEMA
+  directive @override(from: String!) on FIELD_DEFINITION
+  directive @requiresScopes(scopes: [[openfed__Scope!]!]!) on ENUM | FIELD_DEFINITION | INTERFACE | OBJECT | SCALAR
+  directive @shareable on FIELD_DEFINITION | OBJECT
+  scalar openfed__FieldSet
+  scalar openfed__Scope
+`;