+++ connect/src/wg/cosmo/node/v1/node_pb.ts
@@ -1394,9 +1394,9 @@ export class EventConfiguration extends Message<EventConfiguration> {
   topic = "";
 
   /**
-   * @generated from field: string source_id = 5;
+   * @generated from field: string source_name = 5;
    */
-  sourceId = "";
+  sourceName = "";
 
   constructor(data?: PartialMessage<EventConfiguration>) {
     super();
@@ -1410,7 +1410,7 @@ export class EventConfiguration extends Message<EventConfiguration> {
     { no: 2, name: "type_name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
     { no: 3, name: "field_name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
     { no: 4, name: "topic", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 5, name: "source_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 5, name: "source_name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
   ]);
 
   static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): EventConfiguration {